Constructors

Constructors

Constructing MwlsObject directly is not recommended and helper functions described below should be used instead.

MovingWeightedLeastSquares.mwls_cll — Function.

mwls_cll(inputs::Array{T, N}, outputs::Array{U}, EPS::Real, weightfunc::Function) where {T <: Real, U <: Real, N}
mwls_cll(inputs::Array{T, N}, outputs::Array{U}, EPS::Real, weightfunc::Function; maxDegree::Int = 2) where {T <: Real, U <: Real, N}

Creates MwlsCllObject from sample input and sample output data, the cutoff distance ε and a weighting function θ.

Arguments

inputs: a 2d array of input points where each point is on a single row,

outputs: a 2d array or a vector of output scalars where each output is on a single row,

EPS::Real: ε of the method (cell edge length and the default distance for range search),

weightfunc::Function: weighting function θ of the method. It should be in form (distance between two vectors, EPS) -> Float64.

Keyword arguments

maxDegree::Int: the maximal degree of polynomials used for approximation, 2 by default.

source

mwls_cll(input::Array{T, 2}, EPS::Real, weightfunc::Function) where {T <: Real}
mwls_cll(input::Array{T, 2}, EPS::Real, weightfunc::Function; outputDim::Int = 1, maxDegree::Int = 2) where {T <: Real}

In this mwls_cll function, the sample input and sample output data are passed in a single array. It is assumed that each pair of input and output is on a single row. Dimension of the output is specified with kwarg outputDim.

source

MovingWeightedLeastSquares.mwls_kd — Function.

mwls_kd(inputs::Array{T, N}, outputs::Array{U}, EPS::Real, weightfunc::Function) where {T <: Real, U <: Real, N}
mwls_kd(inputs::Array{T, N}, outputs::Array{U}, EPS::Real, weightfunc::Function; leafsize::Int = 10, maxDegree::Int = 2) where {T <: Real, U <: Real, N}

Creates MwlsKdObject from sample input and sample output data, the cutoff distance ε and a weighting function θ.

Arguments

inputs: a 2d array or a vector of input points where each point is on a single row,

outputs: a 2d array or a vector of output scalars where each output is on a single row,

EPS::Real: ε of the method (default distance threshold for neighbor search),

weightfunc::Function: weighting function of the method. It should be in form (distance, EPS) -> Float64.

Keyword arguments

leafSize::Int: the size of the leaves in the k-d-tree, 10 by default.

maxDegree::Int: the maximal degree of polynomials used for approximation, 2 by default.

source

mwls_kd(input::Array{T, 2}, EPS::Real, weightfunc::Function) where {T <: Real}
mwls_kd(input::Array{T, 2}, EPS::Real, weightfunc::Function; outputDim::Int = 1, leafSize::Int = 10, maxDegree::Int = 2) where {T <: Real}

In this mwls_kd function, the sample input and sample output data are passed in a single array. It is assumed that each pair of input and output is on a single row. Dimension of the output is specified with kwarg outputDim.

source